extract shared CopyButton, move invite button to members section, default role to member
